Move BillingDetailsCollectionConfiguration into PaymentElement.Configuration.
